#156 Created new directories for elements and layouts, moved associated files and update @import urls
