+++ src/elements/e-label.vue
