+++ src/setup/styleguide.routes.js
@@ -1,4 +1,4 @@
-import lDefault from '@/components/l-default';
+import lDefault from '@/layouts/l-default.vue';
 
 import index from '@/styleguide/routes/index';
 import sandbox from '@/styleguide/routes/sandbox';
